Use INSTAPAPER_CONSUMER_* environment variables
This moves the API consumer secrets out of the source code and into the
build environment. They can also be added to a local `.env` file, and it
will be read at build time. The values will be expanded directly into
the `main.js` file as part of the build.