Await connection and disconnection promises
In particular, failing to wait for disconnectAccount() to complete in
the error path meant that there could be a race between resetting the
account state and redisplaying the Settings view, which now depends on
the account state to determine whether or not to show most settings.

Also, change the modal's "Connect" button to "Connecting ..." while a
connection attempt is in progress to better communicate the asynchronous
state to the user.