Generalize the sync operation
This was originally written assuming the sync was notes-specific, and
while that's still the only thing we sync with the server, it's a good
time to generalize this code in preparation for the future.

Part of this change renames some of our settings keys. I added some
migration logic to map the old keys to the new ones. We can remove this
in a few versions after we give users some time to upgrade.