Add block identifiers to highlights (#159)
Obsidian uses a `#^` Markdown link extension format to uniquely identify
individual blocks. This also facilitates cross-note linking.

https://help.obsidian.md/links#Link+to+a+block+in+a+note

This approach is more Obsidian-native than our link-based identifiers,
which we used for detecting existing highlights in a node.

Note that we're not exactly following the Obsidian convention of placing
the block identifier _following_ the block quote because that would
interfere with the placement of our user notes. I think that's a good
compromise for our use case, but it's also something we can revisit.